Added process definition read, read_instance & update_instance permissions
Due to the enforcement of Camunda permissions the process definition is now mandatory for a user to successfully submit a job result. The other two were added as a precaution.

Note: This is a potential **breaking change** for users not using the `scanner` group provided by default for their scanner user permissions.
The permissions of the default `scanner` group will get updated by default, all others will have to manually expand their scanner groups to match these permissions.